GCPDWH/workday/requirements.txt
google-cloud-bigquery==1.3.0
google-cloud-storage==1.10.0